+++ packages/core/CHANGELOG.md
@@ -9,6 +9,7 @@ Versioning].
 ## [Unreleased]
 
 - Fix counting errors of attributes in HTML to improve compression.
+- Fix incorrect handling of missing groups in `NestedGroupMangleExpression`.
 - Fix matching of query selectors without prefix.
 
 ## [0.1.21] - 2021-05-30